+++ ExtractReadableValue.ts
@@ -1,7 +1,7 @@
 import type { Readable } from './lib'
-export type extract_readable_val_type<I extends Readable<unknown>> =
+export type ExtractReadableValue<I extends Readable<unknown>, E extends unknown = unknown> =
 	I extends (Readable<infer O>)
 	? O
 	: Extract<I, Readable<unknown>> extends Readable<infer O>
 		? O
-		: unknown
+		: E