+++ ExtractStoresValues.ts
@@ -1,0 +1,14 @@
+import type { Stores } from './Stores'
+import type { Readable } from './lib'
+export type ExtractStoresValues<I extends Stores, E extends unknown = unknown> =
+  I extends (Readable<infer O>)
+    ? O
+    : {
+      [K in keyof I]: (
+        I[K] extends (Readable<infer O>)
+          ? O
+          : Extract<I[K], Readable<unknown>> extends Readable<infer O>
+          ? O
+          : E
+        );
+    }