+++ extract_readable_val_type.d.ts
@@ -1,2 +1,0 @@
-import type { Readable } from './lib';
-export declare type extract_readable_val_type<I extends Readable<unknown>> = I extends (Readable<infer O>) ? O : Extract<I, Readable<unknown>> extends Readable<infer O> ? O : unknown;