+++ lib.d.ts
@@ -1,2 +1,2 @@
-export { derived, writable, readable, } from 'svelte/store';
+export { writable, readable, } from 'svelte/store';
 export type { Writable, Readable } from 'svelte/store';