+++ spread_derived.d.ts
@@ -1,4 +1,4 @@
-import { Readable } from './lib';
+import type { Readable } from './lib';
 import type { Stores } from './Stores';
 export declare function spread_derived<I extends Stores<unknown>, O extends unknown>(stores: I, fn: any, initial_value?: any): Readable<O>;
 export declare const derived__spread: typeof spread_derived;