+++ spread_derived.ts
@@ -1,8 +1,9 @@
-import { _spread } from '@ctx-core/array'
-import { derived, Readable } from './lib'
+import type { Readable } from './lib'
 import type { Stores } from './Stores'
+import { _spread } from '@ctx-core/array'
+import { derived } from './derived'
 export function spread_derived<I extends Stores<unknown>, O extends unknown>(
-	stores: I, fn, initial_value?
+	stores:I, fn, initial_value?
 ) {
 	return derived<I, O>(stores, _spread(fn), initial_value) as Readable<O>
 }